(edit) use DatasetActionButton in SearchCardAction, adjust SearchCard JSX (create SearchCard.js and make index.js for exporting shared sub-components for clean imports), and clean up temporary code blocks for demo
